+++ zingolib/src/lightclient/error.rs
@@ -1,16 +1,9 @@
 //! Errors assoicated with [`crate::lightclient::LightClient`].
 
-use std::convert::Infallible;
-
-use zcash_protocol::TxId;
-
 use pepper_sync::error::{SyncError, SyncModeError};
 use zingo_netutils::GetClientError;
 
-use crate::wallet::{
-    error::{CalculateTransactionError, ProposeSendError, ProposeShieldError, WalletError},
-    output::OutputRef,
-};
+use crate::wallet::error::WalletError;
 
 #[derive(Debug, thiserror::Error)]
 pub enum LightClientError {
@@ -26,9 +19,6 @@ pub enum LightClientError {
     /// Sync mode error.
     #[error("Sync mode error. {0}")]
     SyncModeError(#[from] SyncModeError),
-    /// Send error.
-    #[error("Send error. {0}")]
-    SendError(#[from] SendError),
     /// gPRC client error.
     #[error("gRPC client error. {0}")]
     ClientError(#[from] GetClientError),
@@ -38,41 +28,4 @@ pub enum LightClientError {
     /// Wallet error.
     #[error("Wallet error. {0}")]
     WalletError(#[from] WalletError),
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum SendError {
-    /// Propose send error.
-    #[error("Propose send error. {0}")]
-    ProposeSendError(#[from] ProposeSendError),
-    /// Propose shield error.
-    #[error("Propose shield error. {0}")]
-    ProposeShieldError(#[from] ProposeShieldError),
-    /// Failed to construct sending transaction.
-    #[error("Failed to construct sending transaction. {0}")]
-    CalculateSendError(CalculateTransactionError<OutputRef>),
-    /// Failed to construct shielding transaction.
-    #[error("Failed to construct shielding transaction. {0}")]
-    CalculateShieldError(CalculateTransactionError<Infallible>),
-    /// No proposal found in the wallet.
-    #[error("No proposal found in the wallet.")]
-    NoStoredProposal,
-    /// Transmission error.
-    #[error("Transmission error. {0}")]
-    TransmissionError(#[from] TransmissionError),
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum TransmissionError {
-    /// Transmission failed.
-    #[error("Transmission failed. {0}")]
-    TransmissionFailed(String),
-    /// Transaction to transmit does not have `Calculated` status: {0}
-    #[error("Transaction to transmit does not have `Calculated` status: {0}")]
-    IncorrectTransactionStatus(TxId),
-    /// Txid reported by server does not match calculated txid.
-    #[error(
-        "Server error: txid reported by the server does not match calculated txid.\ncalculated txid:\n{0}\ntxid from server: {1}"
-    )]
-    IncorrectTxidFromServer(TxId, TxId),
 }