+++ zingolib/src/testutils.rs
@@ -31,6 +31,7 @@ pub mod macros;
 pub mod mock_indexer;
 pub mod paths;
 pub mod scenarios;
+pub mod send;
 pub mod synthetic_wallet;
 
 // Re-export test dependencies for convenience